feat(core): split up the functions across diff files
refs. metriport/metriport-internal#2569
